process: extract
